lock-in/IEP1/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
openai>=1.3.0
python-dotenv==0.19.0
httpx>=0.24.1 
prometheus-flask-exporter==0.22.4
prometheus-client==0.17.1
psutil==5.9.5 
